Convert files from JS to TS and edit
