+++ package.json
@@ -13,10 +13,13 @@
   "dependencies": {
     "@chakra-ui/icons": "^2.0.19",
     "@chakra-ui/react": "^2.6.1",
+    "@cypress/code-coverage": "^3.12.44",
+    "@cypress/webpack-preprocessor": "^6.0.2",
     "@emotion/react": "^11.11.0",
     "@emotion/styled": "^11.11.0",
     "@reduxjs/toolkit": "^1.9.5",
     "axios": "^1.4.0",
+    "cypress": "^13.13.2",
     "framer-motion": "^10.12.10",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
@@ -27,7 +30,8 @@
     "react-router-dom": "^6.11.1",
     "react-scripts": "^5.0.1",
     "redux-logger": "^3.0.6",
-    "universal-cookie": "^4.0.4"
+    "universal-cookie": "^4.0.4",
+    "vite-plugin-istanbul": "^6.0.2"
   },
   "devDependencies": {
     "@types/react": "^18.0.28",
