+++ tailwind.config.js
@@ -1,0 +1,5 @@
+/** @type {import('tailwindcss').Config} */
+export default {
+	content: ['./index.html', './src/**/*.{js,ts,jsx,tsx}'],
+	plugins: [],
+};
